requirements.txt
cycler==0.11.0
fonttools==4.28.5
joblib==1.1.0
kiwisolver==1.3.2
numpy
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
six==1.16.0
threadpoolctl==3.0.0
